Adjust set value id
