+++ app/src/main/AndroidManifest.xml
@@ -11,23 +11,23 @@
         android:theme="@style/Theme.BagVana"
         tools:targetApi="31">
         <activity
-            android:name=".LoginRegister.ChangePassword"
+            android:name=".Activity.LoginRegister.ChangePassword"
             android:exported="false" />
         <activity
-            android:name=".LoginRegister.OTPActivity"
+            android:name=".Activity.LoginRegister.OTPActivity"
             android:exported="true" />
         <activity
-            android:name=".LoginRegister.ForgotPassword"
+            android:name=".Activity.LoginRegister.ForgotPassword"
             android:exported="true"/>
 
 
 
         <activity
-            android:name=".LoginRegister.SignUpActivity"
+            android:name=".Activity.LoginRegister.SignUpActivity"
             android:exported="true"/>
 
         <activity
-            android:name=".LoginRegister.SignInActivity"
+            android:name=".Activity.LoginRegister.SignInActivity"
             android:exported="true">
             <intent-filter>
                 <action android:name="android.intent.action.MAIN" />
