+++ app/src/main/java/com/example/bagvana/Activity/Dialog/Dialog.java
@@ -1,0 +1,5 @@
+package com.example.bagvana.Activity.Dialog;
+
+public class Dialog {
+
+}
